Перенос логики отрисовки фильмов в отдельный компонент
