+++ src/const.js
@@ -1,3 +1,5 @@
+export const START_INDEX = 0;
+
 export const RenderPosition = {
   AFTERBEGIN: `afterbegin`,
   BEFOREEND: `beforeend`,
@@ -15,13 +17,6 @@ export const MoviesFilter = {
   FAVORITES: `favorites`,
 };
 
-export const filterToNavItemName = {
-  [MoviesFilter.ALL]: `All movies`,
-  [MoviesFilter.WATCHLIST]: `Watchlist`,
-  [MoviesFilter.HISTORY]: `History`,
-  [MoviesFilter.FAVORITES]: `Favorites`,
-};
-
 export const ExtraTitle = {
   TOP_RATED: `Top rated`,
   TOP_COMMENTED: `Top commented`,
@@ -36,6 +31,13 @@ export const UserRating = {
 export const RatingStep = {
   FIRST: 10,
   SECOND: 20,
+};
+
+export const filterToNavItemName = {
+  [MoviesFilter.ALL]: `All movies`,
+  [MoviesFilter.WATCHLIST]: `Watchlist`,
+  [MoviesFilter.HISTORY]: `History`,
+  [MoviesFilter.FAVORITES]: `Favorites`,
 };
 
 export const EMOJIS = [
